requirements.txt
astroid==2.3.3
click==7.1.1
Flask==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.4
pylint==2.4.4
six==1.14.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.11.2
